Bugfixes
-Fixed mirror texture wrapping detection in renderer
-Fixed renderer TriCount and VertCount max with uint rather than short
-Added functions class (Just a start for now)
-Various other small adjustments

+++ Renderer.cs
@@ -27,8 +27,8 @@ class Renderer
 
     public static Camera cam = new Camera();
     public static Vector2 lastMousePos = new Vector2();
-    public static short TriCount = 0;
-    public static short VertexCount = 0;
+    public static uint TriCount = 0;
+    public static uint VertexCount = 0;
 
     public static void InitialiseView()
     {
